remove FsPartSource
